+++ lib/prody/sequence/__init__.py
@@ -62,6 +62,3 @@ from . import plotting
 from .plotting import *
 __all__.extend(plotting.__all__)
 
-from . import functions
-from .functions import *
-__all__.extend(functions.__all__)